refactor: Complete NWS wrapper modularization (690→4 focused modules)
• Refactored monolithic 690-line nws_wrapper.py into 4 specialized modules:
  - core_client.py (108 lines): HTTP client & request handling
  - point_location.py (186 lines): Geographic data & location identification
  - weather_data.py (284 lines): Current conditions & forecasts
  - alerts_discussions.py (309 lines): Alerts & national products
  - __init__.py (134 lines): Main coordinator with backward compatibility

• Maintained complete backward compatibility:
  - Exact same public API interface preserved
  - All 28 API wrapper tests passing
  - Updated import path to accessiweather.api.nws.NwsApiWrapper

• Improved maintainability:
  - Clear separation of concerns by functionality
  - All modules under 500-line limit (most under 300)
  - Dependency injection pattern for module coordination
  - Consistent error handling and logging

• Clean integration:
  - Updated all test import paths
  - Removed backup files
  - No breaking changes to downstream components

Task: Complete systematic file length refactoring following established patterns
