+++ tests/api_wrapper/test_caching.py
@@ -26,7 +26,7 @@ SAMPLE_POINT_DATA = {
 def api_wrapper():
     """Create a NoaaApiWrapper instance without caching."""
     with (
-        patch("accessiweather.api.nws_wrapper.NwsApiWrapper") as mock_nws,
+        patch("accessiweather.api.nws.NwsApiWrapper") as mock_nws,
         patch("accessiweather.api.openmeteo_wrapper.OpenMeteoApiWrapper") as mock_openmeteo,
     ):
 
@@ -57,7 +57,7 @@ def api_wrapper():
 def cached_api_wrapper():
     """Create a NoaaApiWrapper instance with caching enabled."""
     with (
-        patch("accessiweather.api.nws_wrapper.NwsApiWrapper") as mock_nws,
+        patch("accessiweather.api.nws.NwsApiWrapper") as mock_nws,
         patch("accessiweather.api.openmeteo_wrapper.OpenMeteoApiWrapper") as mock_openmeteo,
     ):
