+++ tests/api_wrapper/test_initialization.py
@@ -70,7 +70,7 @@ SAMPLE_POINT_DATA = {
 def api_wrapper():
     """Create a NoaaApiWrapper instance without caching."""
     with (
-        patch("accessiweather.api.nws_wrapper.NwsApiWrapper") as mock_nws,
+        patch("accessiweather.api.nws.NwsApiWrapper") as mock_nws,
         patch("accessiweather.api.openmeteo_wrapper.OpenMeteoApiWrapper") as mock_openmeteo,
     ):
 
@@ -111,7 +111,7 @@ def api_wrapper():
 def cached_api_wrapper():
     """Create a NoaaApiWrapper instance with caching enabled."""
     with (
-        patch("accessiweather.api.nws_wrapper.NwsApiWrapper") as mock_nws,
+        patch("accessiweather.api.nws.NwsApiWrapper") as mock_nws,
         patch("accessiweather.api.openmeteo_wrapper.OpenMeteoApiWrapper") as mock_openmeteo,
     ):
 
@@ -187,7 +187,7 @@ def test_init_with_caching():
 def test_generate_cache_key():
     """Test cache key generation through NWS wrapper."""
     with (
-        patch("accessiweather.api.nws_wrapper.NwsApiWrapper") as mock_nws,
+        patch("accessiweather.api.nws.NwsApiWrapper") as mock_nws,
         patch("accessiweather.api.openmeteo_wrapper.OpenMeteoApiWrapper") as mock_openmeteo,
     ):
