+++ tests/api_wrapper/test_rate_limiting.py
@@ -26,7 +26,7 @@ SAMPLE_POINT_DATA = {
 def api_wrapper():
     """Create a NoaaApiWrapper instance without caching."""
     with (
-        patch("accessiweather.api.nws_wrapper.NwsApiWrapper") as mock_nws,
+        patch("accessiweather.api.nws.NwsApiWrapper") as mock_nws,
         patch("accessiweather.api.openmeteo_wrapper.OpenMeteoApiWrapper") as mock_openmeteo,
     ):
